[FEAT] - Tractores se pueden seleccionar, refactorizar agregarGrafica
