+++ harvester-app/src/framework/utils/scripts/paginas/analisis/formulas/eliminarCuadroFormulas.js
@@ -1,0 +1,14 @@
+/**
+ * Elimina el cuadro de fórmulas si existe.
+ * @returns {void}
+ */
+function eliminarCuadroFormulas() {
+  const cuadroFormulasExistente = document.querySelector('.contenedor-formulas');
+  if (cuadroFormulasExistente) {
+    cuadroFormulasExistente.remove();
+  }
+}
+
+module.exports = {
+    eliminarCuadroFormulas
+};
