+++ harvester-app/src/framework/utils/scripts/paginas/analisis/graficas/actualizarGraficaConColumna.js
@@ -1,0 +1,122 @@
+const Chart = require('chart.js/auto');
+const ChartDataLabels = require('chartjs-plugin-datalabels');
+const { procesarDatosUniversal } = require(`${rutaBase}/src/framework/utils/scripts/paginas/analisis/graficas/procesarDatosUniversal.js`);
+Chart.register(ChartDataLabels);
+
+/**
+ * Actualiza la gráfica con los datos de una columna específica.
+ * @param {number} graficaId - ID de la gráfica a actualizar.
+ * @param {string} nombreColumna - Nombre de la columna seleccionada.
+ * @returns {void}
+ */
+function actualizarGraficaConColumna(graficaId, nombreColumna, datosOriginalesFormulas) {
+  // Obtener la gráfica
+  const graficaDiv = document.getElementById(`previsualizacion-grafica-${graficaId}`);
+  if (!graficaDiv) {
+    console.error('No se encontró la gráfica con ID:', graficaId);
+    return;
+  }
+
+  const canvas = graficaDiv.querySelector('canvas');
+  if (!canvas) {
+    console.error('No se encontró el canvas de la gráfica');
+    return;
+  }
+
+  const contexto = canvas.getContext('2d');
+  const graficaExistente = Chart.getChart(contexto);
+  
+  if (!graficaExistente) {
+    console.error('No se encontró la instancia de Chart.js');
+    return;
+  }
+
+  // Obtener la hoja seleccionada del localStorage
+  const hojaSeleccionada = localStorage.getItem('hojaSeleccionada');
+  const datos = localStorage.getItem('datosExcel');
+  
+  if (!datos) {
+    mostrarAlerta('Error', 'No hay datos cargados para mostrar en la gráfica.', 'error');
+    return;
+  }
+
+  try {
+    let datosHoja = null;
+    
+    // Parsear los datos del localStorage
+    const datosParseados = JSON.parse(datos);
+    
+    // Determinar qué hoja usar
+    if (hojaSeleccionada && hojaSeleccionada.trim() !== '') {
+      // Usar la hoja seleccionada específica
+      if (datosParseados.hojas && datosParseados.hojas[hojaSeleccionada]) {
+        datosHoja = datosParseados.hojas[hojaSeleccionada];
+      } else {
+        mostrarAlerta('Error', `No se encontró la hoja "${hojaSeleccionada}" en los datos.`, 'error');
+        return;
+      }
+    } else {
+      // Fallback: usar la primera hoja disponible o los datos directos
+      if (datosParseados.hojas) {
+        const nombrePrimeraHoja = Object.keys(datosParseados.hojas)[0];
+        datosHoja = datosParseados.hojas[nombrePrimeraHoja];
+      } else if (Array.isArray(datosParseados)) {
+        datosHoja = datosParseados;
+      } else {
+        mostrarAlerta('Error', 'No se pudieron cargar los datos de la hoja.', 'error');
+        return;
+      }
+    }
+
+    if (!datosHoja || datosHoja.length === 0) {
+      mostrarAlerta('Error', 'La hoja seleccionada no contiene datos.', 'error');
+      return;
+    }
+
+    // Encontrar el índice de la columna
+    const encabezados = datosHoja[0];
+    const indiceColumna = encabezados.indexOf(nombreColumna);
+    
+    if (indiceColumna === -1) {
+      mostrarAlerta('Error', `No se encontró la columna "${nombreColumna}" en la hoja "${hojaSeleccionada || 'seleccionada'}".`, 'error');
+      return;
+    }
+
+    // Extraer los datos de la columna (omitiendo el encabezado)
+    const datosColumna = datosHoja.slice(1).map(fila => fila[indiceColumna]);
+    
+    // GUARDAR DATOS ORIGINALES DE LA COLUMNA
+    datosOriginalesFormulas.set(graficaId, {
+      datos: datosColumna,
+      nombre: nombreColumna,
+      tipo: 'columna',
+      hoja: hojaSeleccionada || 'Hoja por defecto'
+    });
+
+    const tipoGraficaActual = graficaExistente.config.type;
+    
+    // Usar el procesamiento universal
+    const datosRebuild = procesarDatosUniversal(datosColumna, tipoGraficaActual, nombreColumna);
+
+    // Actualizar la gráfica
+    const tituloHoja = hojaSeleccionada ? ` (${hojaSeleccionada})` : '';
+    graficaExistente.options.plugins.title.text = `Datos de: ${nombreColumna}${tituloHoja}`;
+    graficaExistente.data.labels = datosRebuild.labels;
+    graficaExistente.data.datasets[0].data = datosRebuild.valores;
+    graficaExistente.data.datasets[0].label = nombreColumna;
+    
+    // Configurar etiquetas: ocultar SOLO en gráficas de línea
+    const tipoActual = graficaExistente.config.type;
+    graficaExistente.options.plugins.datalabels.display = tipoActual !== 'line';
+    
+    graficaExistente.update();
+
+  } catch (error) {
+    console.error('Error al procesar los datos de la columna:', error);
+    mostrarAlerta('Error', 'Error al procesar los datos de la columna seleccionada.', 'error');
+  }
+}
+
+module.exports = {
+    actualizarGraficaConColumna
+};
